# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Emerald
  description: Memory and context infrastructure for AI agents
  version: 0.3.0
  contact:
    name: Emerald
    url: https://emerald.ai
  license:
    name: MIT
paths:
...
  /v1/search:
    post:
      tags:
      - Search
      summary: Search
      description: Hybrid search across memory (graph) and RAG (vector).
      operationId: search_v1_search_post
      parameters:
      - name: page_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Page token for pagination
          title: Page Token
        description: Page token for pagination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search V1 Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Search
      summary: Search Get
      description: GET variant of search.
      operationId: search_get_v1_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          title: Q
      - name: entity_id
        in: query
        required: true
        schema:
          type: string
          title: Entity Id
      - name: search_mode
        in: query
        required: false
        schema:
          type: string
          default: hybrid
          title: Search Mode
      - name: top_k
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 30
          title: Top K
      - name: rewrite_query
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rewrite Query
      - name: min_confidence
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Min Confidence
      - name: dynamic_truncation
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Dynamic Truncation
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Get V1 Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v2/search:
    post:
      tags:
      - Search
      summary: Search
      description: Hybrid search across memory (graph) and RAG (vector).
      operationId: search_v2_search_post
      parameters:
      - name: page_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Page token for pagination
          title: Page Token
        description: Page token for pagination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search V2 Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Search
      summary: Search Get
      description: GET variant of search.
      operationId: search_get_v2_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          title: Q
      - name: entity_id
        in: query
        required: true
        schema:
          type: string
          title: Entity Id
      - name: search_mode
        in: query
        required: false
        schema:
          type: string
          default: hybrid
          title: Search Mode
      - name: top_k
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 30
          title: Top K
      - name: rewrite_query
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rewrite Query
      - name: min_confidence
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Min Confidence
      - name: dynamic_truncation
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Dynamic Truncation
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Get V2 Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SearchRequest:
      properties:
        q:
          type: string
          title: Q
          examples:
          - 用户偏好什么编程语言？
        entity_id:
          type: string
          title: Entity Id
          examples:
          - user_123
        search_mode:
          type: string
          title: Search Mode
          default: hybrid
          examples:
          - hybrid
        top_k:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Top K
          default: 30
        rerank:
          type: boolean
          title: Rerank
          default: false
        rewrite_query:
          type: boolean
          title: Rewrite Query
          default: false
        filters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Filters
        min_confidence:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Min Confidence
        dynamic_truncation:
          type: boolean
          title: Dynamic Truncation
          default: true
      type: object
      required:
      - q
      - entity_id
      title: SearchRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: em_<32-hex>
      description: All endpoints require a Bearer API key in the `Authorization` header. Keys are prefixed
        with `em_` and scoped to a single entity. The server stores only the SHA-256 hash.
servers:
- url: http://localhost:8000/v1
  description: Local development
- url: https://api.emerald.ai/v1
  description: Production
...
security:
- ApiKeyAuth: []